requirements.txt
asgiref==3.6.0
autopep8==2.0.1
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
dateparser==1.1.7
distlib==0.3.6
Django==4.1.5
djangorestframework==3.14.0
filelock==3.9.0
idna==3.4
mysql-connector-python==8.0.32
platformdirs==2.6.2
protobuf==3.20.3
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
virtualenv==20.17.1
